Added Logger class to log algorithm behaviour for improvements purposes

+++ Knapsack_DP/Knapsack_DP/Timer.cpp
@@ -29,7 +29,7 @@ double Timer::getTime(Resolution resolution)
 	elapsedMicroseconds.QuadPart *= 1000000;
 	elapsedMicroseconds.QuadPart /= frequency.QuadPart;
 
-	double retVal = elapsedMicroseconds.QuadPart;
+	double retVal = (double)elapsedMicroseconds.QuadPart;
 	switch (resolution)
 	{
 		case MICROSECONDS:
